requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.11.2
dj-database-url==2.2.0
drf-spectacular==0.27.2
whitenoise==6.7.0
gunicorn==22.0.0
psycopg[binary]==3.2.1
django-storages[boto3]==1.14.4
boto3==1.35.99
Pillow==10.4.0
numpy==2.0.1
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1
